Add navbar & put all styles in global
The navbar works between menu and the beverages page, effectively
switching between them and having a style highlight on the current
tab for visual feedback. The styles of those pages have also been
put all in the global.css file, since by having individual stylesheets
and importing them they where creating problems. This will be applied
to the rest of the pages in the future