Finish clickable functionality on nav & menu
All the navigation features on both the menu and the navbar work as
intended, the nav items highlight according to which page the user is
currently in, nav bar hides when the menu is selected. The styles
have all been compressed to the global.css file. The app itself
works like intended, but there is room for optimisation in the code,
which is what I will be working on in the next commits.